refactor(knowledge-store): split paths.py into a paths/ package
Four concerns become four files behind the same public surface: layout
(on-disk), store_path (the validated /documents value object), naming
(sanitize + allocate), resolve (path -> row), and legacy (the .xml
title derivation). The legacy renderers are now one file to delete at
the cut instead of lines threaded through the rest. No behavior change.
